[chore][pkg/stanza] Cleanup syslog parser operator files
