+++ pkg/stanza/operator/parser/syslog/parser.go
@@ -6,10 +6,8 @@ package syslog // import "github.com/open-telemetry/opentelemetry-collector-cont
 import (
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"regexp"
-	"strings"
 	"time"
 
 	sl "github.com/influxdata/go-syslog/v3"
@@ -17,143 +15,15 @@ import (
 	"github.com/influxdata/go-syslog/v3/octetcounting"
 	"github.com/influxdata/go-syslog/v3/rfc3164"
 	"github.com/influxdata/go-syslog/v3/rfc5424"
-	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/entry"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
-)
-
-const (
-	operatorType = "syslog_parser"
-
-	RFC3164 = "rfc3164"
-	RFC5424 = "rfc5424"
-
-	NULTrailer = "NUL"
-	LFTrailer  = "LF"
 )
 
 var priRegex = regexp.MustCompile(`\<\d{1,3}\>`)
 
-func init() {
-	operator.Register(operatorType, func() operator.Builder { return NewConfig() })
-}
-
-// NewConfig creates a new syslog parser config with default values
-func NewConfig() *Config {
-	return NewConfigWithID(operatorType)
-}
-
-// NewConfigWithID creates a new syslog parser config with default values
-func NewConfigWithID(operatorID string) *Config {
-	return &Config{
-		ParserConfig: helper.NewParserConfig(operatorID, operatorType),
-	}
-}
-
-// Config is the configuration of a syslog parser operator.
-type Config struct {
-	helper.ParserConfig `mapstructure:",squash"`
-	BaseConfig          `mapstructure:",squash"`
-}
-
-// BaseConfig is the detailed configuration of a syslog parser.
-type BaseConfig struct {
-	Protocol                     string  `mapstructure:"protocol,omitempty"`
-	Location                     string  `mapstructure:"location,omitempty"`
-	EnableOctetCounting          bool    `mapstructure:"enable_octet_counting,omitempty"`
-	AllowSkipPriHeader           bool    `mapstructure:"allow_skip_pri_header,omitempty"`
-	NonTransparentFramingTrailer *string `mapstructure:"non_transparent_framing_trailer,omitempty"`
-}
-
-// Build will build a JSON parser operator.
-func (c Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
-	if c.ParserConfig.TimeParser == nil {
-		parseFromField := entry.NewAttributeField("timestamp")
-		c.ParserConfig.TimeParser = &helper.TimeParser{
-			ParseFrom:  &parseFromField,
-			LayoutType: helper.NativeKey,
-		}
-	}
-
-	parserOperator, err := c.ParserConfig.Build(logger)
-	if err != nil {
-		return nil, err
-	}
-
-	proto := strings.ToLower(c.Protocol)
-
-	switch {
-	case proto == "":
-		return nil, fmt.Errorf("missing field 'protocol'")
-	case proto != RFC5424 && (c.NonTransparentFramingTrailer != nil || c.EnableOctetCounting):
-		return nil, errors.New("octet_counting and non_transparent_framing are only compatible with protocol rfc5424")
-	case proto == RFC5424 && (c.NonTransparentFramingTrailer != nil && c.EnableOctetCounting):
-		return nil, errors.New("only one of octet_counting or non_transparent_framing can be enabled")
-	case proto == RFC5424 && c.NonTransparentFramingTrailer != nil:
-		if *c.NonTransparentFramingTrailer != NULTrailer && *c.NonTransparentFramingTrailer != LFTrailer {
-			return nil, fmt.Errorf("invalid non_transparent_framing_trailer '%s'. Must be either 'LF' or 'NUL'", *c.NonTransparentFramingTrailer)
-		}
-	case proto != RFC5424 && proto != RFC3164:
-		return nil, fmt.Errorf("unsupported protocol version: %s", proto)
-	}
-
-	if c.Location == "" {
-		c.Location = "UTC"
-	}
-
-	location, err := time.LoadLocation(c.Location)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load location %s: %w", c.Location, err)
-	}
-
-	return &Parser{
-		ParserOperator:               parserOperator,
-		protocol:                     proto,
-		location:                     location,
-		enableOctetCounting:          c.EnableOctetCounting,
-		allowSkipPriHeader:           c.AllowSkipPriHeader,
-		nonTransparentFramingTrailer: c.NonTransparentFramingTrailer,
-	}, nil
-}
-
 // parseFunc a parseFunc determines how the raw input is to be parsed into a syslog message
 type parseFunc func(input []byte) (sl.Message, error)
-
-func (s *Parser) buildParseFunc() (parseFunc, error) {
-	switch s.protocol {
-	case RFC3164:
-		return func(input []byte) (sl.Message, error) {
-			if s.allowSkipPriHeader && !priRegex.Match(input) {
-				input = append([]byte("<0>"), input...)
-			}
-			return rfc3164.NewMachine(rfc3164.WithLocaleTimezone(s.location)).Parse(input)
-		}, nil
-	case RFC5424:
-		switch {
-		// Octet Counting Parsing RFC6587
-		case s.enableOctetCounting:
-			return newOctetCountingParseFunc(), nil
-		// Non-Transparent-Framing Parsing RFC6587
-		case s.nonTransparentFramingTrailer != nil && *s.nonTransparentFramingTrailer == LFTrailer:
-			return newNonTransparentFramingParseFunc(nontransparent.LF), nil
-		case s.nonTransparentFramingTrailer != nil && *s.nonTransparentFramingTrailer == NULTrailer:
-			return newNonTransparentFramingParseFunc(nontransparent.NUL), nil
-		// Raw RFC5424 parsing
-		default:
-			return func(input []byte) (sl.Message, error) {
-				if s.allowSkipPriHeader && !priRegex.Match(input) {
-					input = append([]byte("<0>"), input...)
-				}
-				return rfc5424.NewMachine().Parse(input)
-			}, nil
-		}
-
-	default:
-		return nil, fmt.Errorf("invalid protocol %s", s.protocol)
-	}
-}
 
 // Parser is an operator that parses syslog.
 type Parser struct {
@@ -166,32 +36,32 @@ type Parser struct {
 }
 
 // Process will parse an entry field as syslog.
-func (s *Parser) Process(ctx context.Context, entry *entry.Entry) error {
+func (p *Parser) Process(ctx context.Context, entry *entry.Entry) error {
 
 	// if pri header is missing and this is an expected behavior then facility and severity values should be skipped.
-	if !s.enableOctetCounting && s.allowSkipPriHeader {
+	if !p.enableOctetCounting && p.allowSkipPriHeader {
 
 		bytes, err := toBytes(entry.Body)
 		if err != nil {
 			return err
 		}
 
-		if s.shouldSkipPriorityValues(bytes) {
-			return s.ParserOperator.ProcessWithCallback(ctx, entry, s.parse, postprocessWithoutPriHeader)
+		if p.shouldSkipPriorityValues(bytes) {
+			return p.ParserOperator.ProcessWithCallback(ctx, entry, p.parse, postprocessWithoutPriHeader)
 		}
 	}
 
-	return s.ParserOperator.ProcessWithCallback(ctx, entry, s.parse, postprocess)
+	return p.ParserOperator.ProcessWithCallback(ctx, entry, p.parse, postprocess)
 }
 
 // parse will parse a value as syslog.
-func (s *Parser) parse(value any) (any, error) {
+func (p *Parser) parse(value any) (any, error) {
 	bytes, err := toBytes(value)
 	if err != nil {
 		return nil, err
 	}
 
-	pFunc, err := s.buildParseFunc()
+	pFunc, err := p.buildParseFunc()
 	if err != nil {
 		return nil, err
 	}
@@ -201,20 +71,54 @@ func (s *Parser) parse(value any) (any, error) {
 		return nil, err
 	}
 
-	skipPriHeaderValues := s.shouldSkipPriorityValues(bytes)
+	skipPriHeaderValues := p.shouldSkipPriorityValues(bytes)
 
 	switch message := slog.(type) {
 	case *rfc3164.SyslogMessage:
-		return s.parseRFC3164(message, skipPriHeaderValues)
+		return p.parseRFC3164(message, skipPriHeaderValues)
 	case *rfc5424.SyslogMessage:
-		return s.parseRFC5424(message, skipPriHeaderValues)
+		return p.parseRFC5424(message, skipPriHeaderValues)
 	default:
 		return nil, fmt.Errorf("parsed value was not rfc3164 or rfc5424 compliant")
 	}
 }
 
-func (s *Parser) shouldSkipPriorityValues(value []byte) bool {
-	if !s.enableOctetCounting && s.allowSkipPriHeader {
+func (p *Parser) buildParseFunc() (parseFunc, error) {
+	switch p.protocol {
+	case RFC3164:
+		return func(input []byte) (sl.Message, error) {
+			if p.allowSkipPriHeader && !priRegex.Match(input) {
+				input = append([]byte("<0>"), input...)
+			}
+			return rfc3164.NewMachine(rfc3164.WithLocaleTimezone(p.location)).Parse(input)
+		}, nil
+	case RFC5424:
+		switch {
+		// Octet Counting Parsing RFC6587
+		case p.enableOctetCounting:
+			return newOctetCountingParseFunc(), nil
+		// Non-Transparent-Framing Parsing RFC6587
+		case p.nonTransparentFramingTrailer != nil && *p.nonTransparentFramingTrailer == LFTrailer:
+			return newNonTransparentFramingParseFunc(nontransparent.LF), nil
+		case p.nonTransparentFramingTrailer != nil && *p.nonTransparentFramingTrailer == NULTrailer:
+			return newNonTransparentFramingParseFunc(nontransparent.NUL), nil
+		// Raw RFC5424 parsing
+		default:
+			return func(input []byte) (sl.Message, error) {
+				if p.allowSkipPriHeader && !priRegex.Match(input) {
+					input = append([]byte("<0>"), input...)
+				}
+				return rfc5424.NewMachine().Parse(input)
+			}, nil
+		}
+
+	default:
+		return nil, fmt.Errorf("invalid protocol %s", p.protocol)
+	}
+}
+
+func (p *Parser) shouldSkipPriorityValues(value []byte) bool {
+	if !p.enableOctetCounting && p.allowSkipPriHeader {
 		// check if entry starts with '<'.
 		// if not it means that the pre header was missing from the body and hence we should skip it.
 		if len(value) > 1 && value[0] != '<' {
@@ -225,7 +129,7 @@ func (s *Parser) shouldSkipPriorityValues(value []byte) bool {
 }
 
 // parseRFC3164 will parse an RFC3164 syslog message.
-func (s *Parser) parseRFC3164(syslogMessage *rfc3164.SyslogMessage, skipPriHeaderValues bool) (map[string]any, error) {
+func (p *Parser) parseRFC3164(syslogMessage *rfc3164.SyslogMessage, skipPriHeaderValues bool) (map[string]any, error) {
 	value := map[string]any{
 		"timestamp": syslogMessage.Timestamp,
 		"hostname":  syslogMessage.Hostname,
@@ -241,11 +145,11 @@ func (s *Parser) parseRFC3164(syslogMessage *rfc3164.SyslogMessage, skipPriHeade
 		value["facility"] = syslogMessage.Facility
 	}
 
-	return s.toSafeMap(value)
+	return p.toSafeMap(value)
 }
 
 // parseRFC5424 will parse an RFC5424 syslog message.
-func (s *Parser) parseRFC5424(syslogMessage *rfc5424.SyslogMessage, skipPriHeaderValues bool) (map[string]any, error) {
+func (p *Parser) parseRFC5424(syslogMessage *rfc5424.SyslogMessage, skipPriHeaderValues bool) (map[string]any, error) {
 	value := map[string]any{
 		"timestamp":       syslogMessage.Timestamp,
 		"hostname":        syslogMessage.Hostname,
@@ -263,11 +167,11 @@ func (s *Parser) parseRFC5424(syslogMessage *rfc5424.SyslogMessage, skipPriHeade
 		value["facility"] = syslogMessage.Facility
 	}
 
-	return s.toSafeMap(value)
+	return p.toSafeMap(value)
 }
 
 // toSafeMap will dereference any pointers on the supplied map.
-func (s *Parser) toSafeMap(message map[string]any) (map[string]any, error) {
+func (p *Parser) toSafeMap(message map[string]any) (map[string]any, error) {
 	for key, val := range message {
 		switch v := val.(type) {
 		case *string:
