+++ pkg/stanza/operator/parser/syslog/parser_test.go
